+++ packages/access-definitions/src/Wrapping/Contracts/Types/SortableTypeInterface.php
@@ -17,7 +17,7 @@ interface SortableTypeInterface extends TypeInterface
      * All properties of this type that can be used to sort corresponding instances.
      *
      * In most use cases this method can return the same array as
-     * {@link ReadableTypeInterface::getReadableProperties()} but you may want to limit
+     * {@link TransferableTypeInterface::getReadableProperties()} but you may want to limit
      * the properties further, e.g. if sorting over some properties is computation heavy or not supported
      * at all. You may also want to allow more properties for sorting than you allowed for reading,
      * but be careful as this may allow guessing values of non-readable properties.