+++ packages/access-definitions/src/Wrapping/Contracts/WrapperFactoryInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace EDT\Wrapping\Contracts;
 
-use EDT\Wrapping\Contracts\Types\ReadableTypeInterface;
+use EDT\Wrapping\Contracts\Types\TransferableTypeInterface;
 use EDT\Wrapping\Contracts\Types\TypeInterface;
 
 /**
@@ -22,12 +22,12 @@ interface WrapperFactoryInterface
     /**
      * @template TEntity of object
      *
-     * @param TEntity                                              $entity
-     * @param ReadableTypeInterface<TCondition, TSorting, TEntity> $type
+     * @param TEntity                                                  $entity
+     * @param TransferableTypeInterface<TCondition, TSorting, TEntity> $type
      *
      * @return array<non-empty-string, mixed>|object|null
      *
      * @throws AccessException
      */
-    public function createWrapper(object $entity, ReadableTypeInterface $type);
+    public function createWrapper(object $entity, TransferableTypeInterface $type);
 }